Split pr_cmd lifecycle into focused modules
